+++ bocoel/corpora/storages/__init__.py
@@ -1,4 +1,4 @@
 from .concat import ConcatStorage
-from .dataframes import DataFrameStorage
 from .datasets import DatasetsStorage
 from .interfaces import Storage
+from .pandas import PandasStorage